fix(acp): clear react-doctor findings on chat module boundaries
Split helpers from component modules for Fast Refresh, extract the history row, and load collapse motion via LazyMotion.
