+++ src/renderer/components/chat/TurnActivity.tsx
@@ -8,6 +8,7 @@ import { cn } from '@/lib/utils'
 import { ChatMessage } from './ChatMessage'
 import { CHEVRON_TRANSITION } from './chat-motion'
 import type { TimelineItem } from './chat-timeline'
+import { formatTurnDuration } from './format-turn-duration'
 import { ThoughtGroup } from './ThoughtGroup'
 import { ToolCallCard } from './ToolCallCard'
 
@@ -18,18 +19,6 @@ interface TurnActivityProps {
   attentionRequired: boolean
   hasFinalResponse: boolean
   shouldAnimateEnter: (id: string) => boolean
-}
-
-/** Whole-second duration without implying precision the available timestamps do not have. */
-function formatTurnDuration(durationMs: number | null): string | null {
-  if (durationMs === null) return null
-  const totalSeconds = Math.max(1, Math.round(durationMs / 1000))
-  const hours = Math.floor(totalSeconds / 3600)
-  const minutes = Math.floor((totalSeconds % 3600) / 60)
-  const seconds = totalSeconds % 60
-  if (hours > 0) return `${hours}h ${minutes}m ${seconds}s`
-  if (minutes > 0) return `${minutes}m ${seconds}s`
-  return `${seconds}s`
 }
 
 /** Borderless, turn-level disclosure for reasoning, tools, and intermediate narration. */
@@ -115,5 +104,3 @@ export function TurnActivity({
     </Collapsible>
   )
 }
-
-export { formatTurnDuration }
