+++ src/renderer/components/chat/format-turn-duration.ts
@@ -1,0 +1,11 @@
+/** Whole-second duration without implying precision the available timestamps do not have. */
+export function formatTurnDuration(durationMs: number | null): string | null {
+  if (durationMs === null) return null
+  const totalSeconds = Math.max(1, Math.round(durationMs / 1000))
+  const hours = Math.floor(totalSeconds / 3600)
+  const minutes = Math.floor((totalSeconds % 3600) / 60)
+  const seconds = totalSeconds % 60
+  if (hours > 0) return `${hours}h ${minutes}m ${seconds}s`
+  if (minutes > 0) return `${minutes}m ${seconds}s`
+  return `${seconds}s`
+}
