+++ src/renderer/components/chat/is-agent-connected.ts
@@ -1,0 +1,8 @@
+import type { AgentStatus } from '@/stores/acp-store'
+
+export function isAgentConnected(
+  session: { status: string } | null | undefined,
+  agentStatus: AgentStatus | undefined
+): boolean {
+  return session != null && session.status !== 'closed' && agentStatus === 'connected'
+}
