+++ src/renderer/components/ui/collapse-expand-motion.tsx
@@ -1,4 +1,4 @@
-import { AnimatePresence, motion, useReducedMotion } from 'framer-motion'
+import { AnimatePresence, domAnimation, LazyMotion, m, useReducedMotion } from 'framer-motion'
 import type { ReactNode } from 'react'
 import { cn } from '@/lib/utils'
 
@@ -6,7 +6,7 @@ import { cn } from '@/lib/utils'
  * Collapse/expand via `grid-template-rows` 0fr→1fr (not `height`), so layout
  * work stays cheaper than animating pixel height. See animations skill.
  */
-export const collapseExpandTransition = {
+const collapseExpandTransition = {
   duration: 0.15,
   ease: 'easeInOut'
 } as const
@@ -27,18 +27,20 @@ export function CollapseExpandMotion({
   const reduced = useReducedMotion() ?? false
 
   return (
-    <AnimatePresence initial={false} onExitComplete={onExitComplete}>
-      {open && (
-        <motion.div
-          initial={reduced ? false : { gridTemplateRows: '0fr', opacity: 0 }}
-          animate={{ gridTemplateRows: '1fr', opacity: 1 }}
-          exit={reduced ? { opacity: 0 } : { gridTemplateRows: '0fr', opacity: 0 }}
-          transition={reduced ? { duration: 0 } : collapseExpandTransition}
-          className={cn('grid overflow-hidden', className)}
-        >
-          <div className="min-h-0 overflow-hidden">{children}</div>
-        </motion.div>
-      )}
-    </AnimatePresence>
+    <LazyMotion features={domAnimation} strict>
+      <AnimatePresence initial={false} onExitComplete={onExitComplete}>
+        {open && (
+          <m.div
+            initial={reduced ? false : { gridTemplateRows: '0fr', opacity: 0 }}
+            animate={{ gridTemplateRows: '1fr', opacity: 1 }}
+            exit={reduced ? { opacity: 0 } : { gridTemplateRows: '0fr', opacity: 0 }}
+            transition={reduced ? { duration: 0 } : collapseExpandTransition}
+            className={cn('grid overflow-hidden', className)}
+          >
+            <div className="min-h-0 overflow-hidden">{children}</div>
+          </m.div>
+        )}
+      </AnimatePresence>
+    </LazyMotion>
   )
 }
