+++ src/renderer/components/workspace/WorkspaceTabBar.tsx
@@ -16,7 +16,8 @@ import { useShallow } from 'zustand/shallow'
 import { AgentIcon } from '@/components/agents/AgentIcon'
 import { ContextMenu } from '@/components/ContextMenu'
 import { AgentBadge } from '@/components/chat/AgentBadge'
-import { AgentConnectionLamp, isAgentConnected } from '@/components/chat/AgentConnectionLamp'
+import { AgentConnectionLamp } from '@/components/chat/AgentConnectionLamp'
+import { isAgentConnected } from '@/components/chat/is-agent-connected'
 import { Skeleton } from '@/components/ui/skeleton'
 import { usePaneDnd } from '@/hooks/use-pane-dnd'
 import { clipboardApi, shellApi } from '@/lib/api'
